renaming data folder to static
